requirements.txt
langchain
langchain-openai
pypdf
faiss-cpu
python-dotenv
tiktoken
langchain-community
streamlit